requirements.txt
streamlit
pdf2image
pytesseract
opencv-python-headless
pandas
numpy
pillow